API: Move lodcm out of the epics folder
